+++ components/forms/company-create-form.tsx
@@ -1,10 +1,6 @@
 "use client"
 
-import {zodResolver} from "@hookform/resolvers/zod"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
 import {CaretSortIcon} from "@radix-ui/react-icons"
-
 import {Button} from "@/components/ui/button"
 import {
   Form,
@@ -15,69 +11,27 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import {Input} from "@/components/ui/input"
-import {
-  usePostMyCompanyMutation,
-  useUpdateMyCompanyMutation,
-} from "@/lib/features/accounts/accountsApiSlice";
 import Loader from "@/components/general/Loader";
 import {cn} from "@/lib/utils";
 import {Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList} from "@/components/ui/command";
 import {Popover, PopoverContent, PopoverTrigger} from "@/components/ui/popover";
 import {Check} from "lucide-react";
 import {countries} from "@/utils/constForm";
-import {toast} from "react-toastify";
 import {Avatar, AvatarFallback, AvatarImage} from "@/components/ui/avatar";
 import getImageData from "@/utils/getImage";
-import {useState} from "react";
 import {Separator} from "@/components/ui/separator";
-import {Company} from "@/lib/features/other/otherApiSlice";
-import {companyFormSchema, CompanyFormValues} from "@/components/forms/company-form";
 import {Textarea} from "@/components/ui/textarea";
+import useCompanyCreateForm from "@/hooks/useCompanyCreateForm";
 
 
 export function CompanyCreateForm() {
-  const [createCompany, {isLoading: isLoadingCreate}] = usePostMyCompanyMutation();
-  const [tempImage, setTempImage] = useState();
-
-  const form = useForm<CompanyFormValues>({
-    resolver: zodResolver(companyFormSchema),
-    defaultValues: {
-      name: '',
-      image: '',
-      bio: '',
-      company_url: '',
-      dou_url: '',
-      country: '',
-      num_employees: 1,
-    },
-    mode: "onChange",
-  })
-
-
-  function onSubmit(data: CompanyFormValues) {
-    console.log('Data: ', data)
-
-    const formData = new FormData();
-    formData.append('name', data.name)
-    formData.append('bio', data.bio)
-    formData.append('company_url', String(data.company_url))
-    formData.append('dou_url', String(data.dou_url))
-    formData.append('country', data.country)
-    formData.append('num_employees', String(data.num_employees))
-    if (typeof data.image === 'object') {
-      formData.append('image', data.image[0], data.image[0].name);
-    }
-
-    // @ts-ignore
-    createCompany(formData)
-      .unwrap()
-      .then(() => {
-        toast.success('Created Company')
-      })
-      .catch(() => {
-        toast.error('Failed to created Company')
-      });
-  }
+  const {
+    form,
+    onSubmit,
+    tempImage,
+    setTempImage,
+    isLoadingCreate,
+  } = useCompanyCreateForm()
 
   return (
     <Form {...form}>
@@ -103,9 +57,8 @@ export function CompanyCreateForm() {
                     className='w-56'
                     onChange={(e) => {
                       const {files, displayUrl} = getImageData(e)
-                      // @ts-ignore
-                      setTempImage(displayUrl)
-                      // @ts-ignore
+
+                      setTempImage(displayUrl);
                       field.onChange(files);
                     }}
                   />
