+++ components/forms/company-form.tsx
@@ -1,10 +1,6 @@
 "use client"
 
-import {zodResolver} from "@hookform/resolvers/zod"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
 import {CaretSortIcon} from "@radix-ui/react-icons"
-
 import {Button} from "@/components/ui/button"
 import {
   Form,
@@ -15,43 +11,18 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import {Input} from "@/components/ui/input"
-import {
-  useUpdateMyCompanyMutation,
-} from "@/lib/features/accounts/accountsApiSlice";
 import Loader from "@/components/general/Loader";
 import {cn} from "@/lib/utils";
 import {Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList} from "@/components/ui/command";
 import {Popover, PopoverContent, PopoverTrigger} from "@/components/ui/popover";
 import {Check} from "lucide-react";
 import {countries} from "@/utils/constForm";
-import {toast} from "react-toastify";
 import {Avatar, AvatarFallback, AvatarImage} from "@/components/ui/avatar";
 import getImageData from "@/utils/getImage";
-import {useState} from "react";
 import {Separator} from "@/components/ui/separator";
 import {Company} from "@/lib/features/other/otherApiSlice";
 import {Textarea} from "@/components/ui/textarea";
-
-export const companyFormSchema = z.object({
-  name: z
-    .string()
-    .min(2, {
-      message: "Company name must be at least 2 characters.",
-    })
-    .max(30, {
-      message: "Company name must not be longer than 30 characters.",
-    }),
-  image: z.any(),
-  bio: z.string(),
-  company_url: z.string().url().optional().or(z.literal('')),
-  dou_url: z.string().url().optional().or(z.literal('')),
-  country: z.string({
-    required_error: "Please select a country.",
-  }),
-  num_employees: z.coerce.number().gte(1, 'Must be 1 or above'),
-})
-
-export type CompanyFormValues = z.infer<typeof companyFormSchema>
+import useCompanyForm from "@/hooks/useCompanyForm";
 
 
 interface CompanyFormProps {
@@ -59,43 +30,13 @@ interface CompanyFormProps {
 }
 
 export function CompanyForm({company}: CompanyFormProps) {
-  const [updateCompany, {isLoading: isLoadingUpdate}] = useUpdateMyCompanyMutation();
-  const [tempImage, setTempImage] = useState(company?.image);
-
-  const form = useForm<CompanyFormValues>({
-    resolver: zodResolver(companyFormSchema),
-    defaultValues: company,
-    mode: "onChange",
-  })
-
-  console.log(company)
-
-
-  function onSubmit(data: CompanyFormValues) {
-    console.log('Data: ', data)
-
-    const formData = new FormData();
-    formData.append('id', String(company?.id))
-    formData.append('name', data.name)
-    formData.append('bio', data.bio)
-    formData.append('company_url', String(data.company_url))
-    formData.append('dou_url', String(data.dou_url))
-    formData.append('country', data.country)
-    formData.append('num_employees', String(data.num_employees))
-    if (typeof data.image === 'object') {
-      formData.append('image', data.image[0], data.image[0].name);
-    }
-
-    // @ts-ignore
-    updateCompany(formData)
-      .unwrap()
-      .then(() => {
-        toast.success('Updated Company')
-      })
-      .catch(() => {
-        toast.error('Failed to update Company')
-      });
-  }
+  const {
+    form,
+    tempImage,
+    setTempImage,
+    onSubmit,
+    isLoadingUpdate,
+  } = useCompanyForm(company)
 
   return (
     <Form {...form}>
@@ -121,9 +62,8 @@ export function CompanyForm({company}: CompanyFormProps) {
                     className='w-56'
                     onChange={(e) => {
                       const {files, displayUrl} = getImageData(e)
-                      // @ts-ignore
+
                       setTempImage(displayUrl)
-                      // @ts-ignore
                       field.onChange(files);
                     }}
                   />
