+++ components/forms/contact-cv-form.tsx
@@ -1,8 +1,5 @@
 "use client"
 
-import {zodResolver} from "@hookform/resolvers/zod"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
 import {Button} from "@/components/ui/button"
 import {
   Form,
@@ -13,49 +10,15 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import {Input} from "@/components/ui/input"
-import {
-  ContactCv, useUpdateMeCandidateContactCvMutation,
-} from "@/lib/features/accounts/accountsApiSlice";
+import {ContactCv} from "@/lib/features/accounts/accountsApiSlice";
 import Loader from "@/components/general/Loader";
 import {PhoneInput} from "@/components/ui/phone-input";
-import {isValidPhoneNumber} from "react-phone-number-input";
 import {File, FileX} from "lucide-react";
 import {Alert, AlertDescription, AlertTitle} from "@/components/ui/alert";
 import {Separator} from "@/components/ui/separator";
 import Link from "next/link";
-import {toast} from "react-toastify";
 import dayjs from "dayjs";
-
-const contactFormSchema = z.object({
-  first_name: z
-    .string()
-    .min(2, {
-      message: "First name must be at least 2 characters.",
-    })
-    .max(30, {
-      message: "First name must not be longer than 30 characters.",
-    }),
-  last_name: z
-    .string()
-    .min(2, {
-      message: "Last name must be at least 2 characters.",
-    })
-    .max(30, {
-      message: "Last name must not be longer than 30 characters.",
-    }),
-  email: z.string().email(),
-  phone_number: z
-    .string()
-    .refine(isValidPhoneNumber, {message: "Invalid phone number"})
-    .or(z.literal("")),
-  telegram_url: z.string().url().optional().or(z.literal('')),
-  linkedin_url: z.string().url().optional().or(z.literal('')),
-  git_hub_url: z.string().url().optional().or(z.literal('')),
-  portfolio_url: z.string().url().optional().or(z.literal('')),
-  cv_file: z.any(),
-})
-
-type ContactFormValues = z.infer<typeof contactFormSchema>
+import useContactCvForm from "@/hooks/useContactCvForm";
 
 
 interface ContactFormProps {
@@ -63,46 +26,11 @@ interface ContactFormProps {
 }
 
 export function ContactCvForm({contactCv}: ContactFormProps) {
-  const [updateCandidateCv, {isLoading: isLoadingUpdate}] = useUpdateMeCandidateContactCvMutation();
-  // @ts-ignore
-
-  const form = useForm<ContactFormValues>({
-    resolver: zodResolver(contactFormSchema),
-    // @ts-ignore
-    defaultValues: contactCv,
-    mode: "onChange",
-  })
-
-  console.log(contactCv)
-
-  function onSubmit(data: ContactFormValues) {
-    console.log('Data: ', data)
-
-    const formData = new FormData();
-    formData.append('first_name', data.first_name)
-    formData.append('last_name', data.last_name)
-    formData.append('email', data.email)
-    formData.append('phone_number', data.phone_number)
-    formData.append('telegram_url', String(data.telegram_url))
-    formData.append('linkedin_url', String(data.linkedin_url))
-    formData.append('git_hub_url', String(data.git_hub_url))
-    formData.append('portfolio_url', String(data.portfolio_url))
-    console.log(typeof data?.cv_file)
-    if (typeof data?.cv_file === "object") {
-      formData.append('cv_file', data.cv_file, data.cv_file.name)
-    }
-
-
-    // @ts-ignore
-    updateCandidateCv(formData)
-      .unwrap()
-      .then(() => {
-        toast.success('Updated ContactCv')
-      })
-      .catch(() => {
-        toast.error('Failed to update ContactCv')
-      });
-  }
+  const {
+    form,
+    onSubmit,
+    isLoadingUpdate,
+  } = useContactCvForm(contactCv)
 
   return (
     <Form {...form}>
@@ -236,7 +164,8 @@ export function ContactCvForm({contactCv}: ContactFormProps) {
               <AlertTitle>
                 {contactCv.cv_file.substring(contactCv.cv_file.lastIndexOf('/') + 1)}
               </AlertTitle>
-              <AlertDescription>Last modify: {dayjs(contactCv.updated_at).format('D MMMM YYYY HH:MM')}</AlertDescription>
+              <AlertDescription>Last
+                modify: {dayjs(contactCv.updated_at).format('D MMMM YYYY HH:MM')}</AlertDescription>
             </Alert>
           </Link>
           : <Alert>
