+++ components/forms/image-form.tsx
@@ -1,54 +1,23 @@
 import {
   Candidate,
-  useUpdateMeCandidateImageMutation
 } from "@/lib/features/accounts/accountsApiSlice";
-import {useState} from "react";
 import {Avatar, AvatarFallback, AvatarImage} from "@/components/ui/avatar";
 import {Form, FormControl, FormField, FormItem, FormMessage} from "@/components/ui/form";
 import {Input} from "@/components/ui/input";
 import getImageData from "@/utils/getImage";
 import {Button} from "@/components/ui/button";
 import Loader from "@/components/general/Loader";
-import {useForm} from "react-hook-form";
-import {zodResolver} from "@hookform/resolvers/zod";
-import {z} from "zod";
-import {toast} from "react-toastify";
+import useImageForm from "@/hooks/useImageForm";
 
-
-const imageFormSchema = z.object({
-  image: z.any()
-})
-
-type ImageFormValues = z.infer<typeof imageFormSchema>
 
 export default function ImageForm({candidate}: { candidate?: Candidate | undefined }) {
-  const [updateImage, {isLoading}] = useUpdateMeCandidateImageMutation();
-  const [tempImage, setTempImage] = useState(candidate?.image);
-
-  const form = useForm<ImageFormValues>({
-    resolver: zodResolver(imageFormSchema),
-    // @ts-ignore
-    defaultValues: candidate?.image,
-    mode: "onChange",
-  })
-
-  function onSubmit(data: ImageFormValues) {
-    if (data?.image) {
-      const formData = new FormData();
-      formData.append('image', data.image[0], data.image[0].name);
-      // @ts-ignore
-      updateImage(formData)
-        .unwrap()
-        .then(() => {
-          toast.success('Updated image')
-        })
-        .catch(() => {
-          toast.error('Failed to update image')
-        });
-    } else {
-      toast.warning('Not Uploaded Image')
-    }
-  }
+  const {
+    form,
+    onSubmit,
+    isLoading,
+    tempImage,
+    setTempImage,
+  } = useImageForm(candidate)
 
   return (
     <Form {...form}>
@@ -72,9 +41,9 @@ export default function ImageForm({candidate}: { candidate?: Candidate | undefin
                   className='w-56'
                   onChange={(e) => {
                     const {files, displayUrl} = getImageData(e)
-                    // @ts-ignore
+
                     setTempImage(displayUrl)
-                    // @ts-ignore
+
                     field.onChange(files);
                   }}
                 />
