+++ components/forms/notifications-form.tsx
@@ -1,9 +1,5 @@
 "use client"
 
-import {zodResolver} from "@hookform/resolvers/zod"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
-
 import {Button} from "@/components/ui/button"
 import {
   Form,
@@ -18,68 +14,22 @@ import {RadioGroup, RadioGroupItem} from "@/components/ui/radio-group"
 import {Switch} from "@/components/ui/switch"
 import {User} from "@/lib/features/auth/authApiSlice";
 import {Checkbox} from "@/components/ui/checkbox";
-import {
-  useDeleteIsSpamEmailEveryWeekMutation,
-  usePostIsSpamEmailEveryWeekMutation
-} from "@/lib/features/accounts/accountsApiSlice";
-import {toast} from "react-toastify";
 import Loader from "@/components/general/Loader";
+import useNotificationsForm from "@/hooks/useNotificationsForm";
 
-const notificationsFormSchema = z.object({
-  type: z.enum(["email", "none"], {
-    required_error: "You need to select a notification type.",
-  }),
-  notify_recruiter_emails: z.boolean().default(false).optional(),
-  social_emails: z.boolean().default(false).optional(),
-  auto_offers: z.boolean().default(false).optional(),
-  is_spam_email: z.boolean().optional(),
-})
-
-type NotificationsFormValues = z.infer<typeof notificationsFormSchema>
 
 interface Props {
   user?: User | undefined
 }
 
 export function NotificationsForm({user}: Props) {
-  const [postSpam, {isLoading: isLoadingPost}] = usePostIsSpamEmailEveryWeekMutation();
-  const [deleteSpam, {isLoading: isLoadingDelete}] = useDeleteIsSpamEmailEveryWeekMutation();
+  const {
+    isLoadingDelete,
+    isLoadingPost,
+    form,
+    onSubmit,
+  } = useNotificationsForm(user)
 
-  const form = useForm<NotificationsFormValues>({
-    resolver: zodResolver(notificationsFormSchema),
-    defaultValues: {
-      type: user?.is_spam_email ? "email" : "none",
-      notify_recruiter_emails: false,
-      social_emails: true,
-      auto_offers: true,
-      is_spam_email: user?.is_spam_email,
-    },
-  })
-
-  function onSubmit(data: NotificationsFormValues) {
-    console.log(data)
-
-    if (data.is_spam_email) {
-      postSpam(undefined)
-        .unwrap()
-        .then(({msg}) => {
-          toast.success(msg)
-        })
-        .catch(() => {
-          toast.error('Failed to update notification')
-        })
-    } else {
-      deleteSpam(undefined)
-        .unwrap()
-        .then(({msg}) => {
-          toast.success(msg)
-        })
-        .catch(() => {
-          toast.error('Failed to update notification')
-        })
-    }
-
-  }
 
   return (
     <Form {...form}>
