+++ components/forms/vacancy-form.tsx
@@ -1,10 +1,6 @@
 "use client"
 
-import {zodResolver} from "@hookform/resolvers/zod"
-import {useForm} from "react-hook-form"
-import {z} from "zod"
 import {CaretSortIcon} from "@radix-ui/react-icons"
-
 import {Button} from "@/components/ui/button"
 import {
   Form,
@@ -15,9 +11,7 @@ import {
   FormMessage,
 } from "@/components/ui/form"
 import {Input} from "@/components/ui/input"
-import {
-  useUpdateMyVacancyMutation, Vacancy,
-} from "@/lib/features/accounts/accountsApiSlice";
+import {Vacancy} from "@/lib/features/accounts/accountsApiSlice";
 import Loader from "@/components/general/Loader";
 import {RadioGroup, RadioGroupItem} from "@/components/ui/radio-group";
 import {Textarea} from "@/components/ui/textarea";
@@ -27,91 +21,28 @@ import {Popover, PopoverContent, PopoverTrigger} from "@/components/ui/popover";
 import {Check} from "lucide-react";
 import {countries, employOptions} from "@/utils/constForm";
 import {Checkbox} from "@/components/ui/checkbox";
-import {useState} from "react";
 import {Tag, TagInput} from "@/components/ui/tag-input";
 import {Category, Company, Skills} from "@/lib/features/other/otherApiSlice";
 import {Slider} from "@/components/ui/slider";
 import {Avatar, AvatarImage} from "@/components/ui/avatar";
-import {toast} from "react-toastify";
-
-export const vacancyFormSchema = z.object({
-  title: z
-    .string()
-    .min(2, {
-      message: "First name must be at least 2 characters.",
-    })
-    .max(100, {
-      message: "First name must not be longer than 100 characters.",
-    }),
-  description: z.string()
-    .max(1000, {message: 'Must not be longer than 1000 characters.',}),
-  requirements: z.string()
-    .max(1000, {message: 'Must not be longer than 1000 characters.',}),
-  other: z.string()
-    .max(1000, {message: 'Must not be longer than 1000 characters.',}),
-  eng_level: z
-    .enum(['none', 'beginner', 'intermediate', 'upper_intermediate', 'advanced']),
-  salary: z.coerce.number().gte(1, 'Must be 1 or above'),
-  category: z
-    .string({required_error: "Category is required"}),
-  skills: z.any(),
-  work_exp: z.coerce.number()
-    .gte(0, 'Must be 0 or above')
-    .lte(10, 'Must be less or equal then 10 '),
-  employ_options: z
-    .array(z.string()).refine((value) => value.some((item) => item)),
-  country: z.string({
-    required_error: "Please select a country.",
-  }),
-  company: z.any({
-    required_error: "Please select a company.",
-  }),
-  is_only_ukraine: z.boolean(),
-  is_test_task: z.boolean(),
-})
-
-export type VacancyFormValues = z.infer<typeof vacancyFormSchema>
+import useVacancyForm from "@/hooks/useVacancyForm";
 
 
 interface VacancyFormProps {
   vacancy?: Vacancy | undefined
-  skills?: Skills | undefined
+  skills: Skills
   category?: Category | undefined
   myCompanies?: Company[] | undefined
 }
 
 export function VacancyForm({vacancy, skills, category, myCompanies}: VacancyFormProps) {
-  const [updateVacancy, {isLoading: isLoadingUpdate}] = useUpdateMyVacancyMutation();
-  // @ts-ignore
-  const [tags, setTags] = useState<Tag[]>(skills?.results.filter((item) => vacancy?.skills.includes(item.text)));
-
-  const form = useForm<VacancyFormValues>({
-    resolver: zodResolver(vacancyFormSchema),
-    // @ts-ignore
-    defaultValues: {
-      ...vacancy,
-      company: vacancy?.company.name
-    },
-    mode: "onChange",
-  })
-
-  console.log(vacancy)
-
-
-  function onSubmit(data: VacancyFormValues) {
-    console.log('Data: ', data)
-
-    console.log(vacancy?.slug)
-    // @ts-ignore
-    updateVacancy({slug: vacancy?.slug, ...data})
-      .unwrap()
-      .then(() => {
-        toast.success('Updated Vacancy')
-      })
-      .catch(() => {
-        toast.error('Failed to update Vacancy')
-      });
-  }
+  const {
+    form,
+    onSubmit,
+    isLoadingUpdate,
+    tags,
+    setTags,
+  } = useVacancyForm(vacancy, skills)
 
   return (
     <Form {...form}>
